requirements.txt
PyQt5==5.15.10
opencv-python==4.9.0.80
Pillow==10.2.0
pydub==0.25.1
soundfile==0.12.1
sounddevice==0.4.6
cryptography==42.0.5
pycryptodome==3.20.0
python-docx==1.1.0
msoffcrypto-tool==5.0.0
mutagen==1.47.0
python-gnupg==0.5.2
ffmpeg-python==0.2.0
numpy==1.26.4
scipy==1.12.0
psutil==5.9.8
stepic==0.5.0
